+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## [0.12.3] - 2024-08-07
 ### Changed
 - Renamed config option `:auto_wire_options` to `:auto_wire` at `:dry_validation` plugin
+- Updated `Pathway::State#unwrap` to raise an `ArgumentError` exception on invalid arguments
+### Added
+- Provide alias `Pathway::State#use` to `Pathway::State#unwrap`
 
 ## [0.12.2] - 2024-08-06
 ### Added